Don't install custom shellcheck/shfmt during test setup
Since they are both present in the base image, and if the cache is
emptied the versions specified in .travis.yml were never installed
anyway, causing confusion over the version number in use. Whilst
this approach will still result in breakage if a new base image version
includes backwards incompatible changes, it saves having to increment
the version in yet another location, plus provides additional testing
of the images themselves.
:tada: :bomb: commit-range=4a3a786b2aaddd6b4d84fd244a7ecf4d7e0e71e3...cb15ef92aa17cc2d0707eeb0baaa1caf7afd52ef